+++ main/gyro.cpp
@@ -10,19 +10,42 @@ Gyro::Gyro() {
     Wire.beginTransmission(MPU);
     Wire.write(0x6B);  
     Wire.write(0);
-    Wire.endTransmission(true); 
+    Wire.endTransmission(true);
 
-    Wire.beginTransmission(MPU);
-    Wire.write(0x3B);  
-    Wire.endTransmission(false);
-    Wire.requestFrom(MPU, 6, true);  // 6 pieces of data cause we don't care about rotational
+    // This sets the initial gyro value to ensure that it is calibrated. Sample size in this case is 500
+    // It will calculate the average gyro magniture over this time, if the value is over x, assume the calibration failed
+    // Like if it was done while accelerating, so set a conservative amount EXPECTED_ACC_MAGNITUDE
+    float sum = 0.0;
+    int n = 0;
+    delay(3000);
+    while (n < CALIBRATION_SAMPLE_SIZE) {
+        Wire.beginTransmission(MPU);
+        Wire.write(0x3B);  
+        Wire.endTransmission(false);
+        Wire.requestFrom(MPU, 6, true);
+        if (Wire.available() == 6) {
+            this->measuredAccX = Wire.read() << 8 | Wire.read();    
+            this->measuredAccY = Wire.read() << 8 | Wire.read();  
+            this->measuredAccZ = Wire.read() << 8 | Wire.read();
+            
+            this->correctedAcc = sqrt(
+                            this->measuredAccX * this->measuredAccX +
+                            this->measuredAccY * this->measuredAccY +
+                            this->measuredAccZ * this->measuredAccZ);
+            n++;
+            sum += this->correctedAcc;
 
-    if (Wire.available() == 6) {
-        Serial.println("Connected to Gyro");
-    } else {
-        Serial.println("Failed to read from MPU");
+        } else {
+            Serial.println("Failed to read from MPU");
+            continue;
+        }
+        delay(5);
     }
-
+    this->idleAcc = sum/(n+1);
+    if (abs(this->idleAcc - EXPECTED_ACC_MAGNITUDE) > CALIBRATION_ACC_DELTA) {
+        this->idleAcc = EXPECTED_ACC_MAGNITUDE;
+    } 
+    
     this->lastUpdateTime = millis(); // Initialize the last update time
 }
 
@@ -42,35 +65,84 @@ void Gyro::Update() {
     }
 
     // Calculate the corrected acceleration
-    this->correctedAcc = sqrt(this->measuredAccX * this->measuredAccX +
-                              this->measuredAccY * this->measuredAccY +
-                              this->measuredAccZ * this->measuredAccZ) - EXPECTED_ACC_MAGNITUDE;
-
-    this->correctedAcc = this->correctedAcc < 0 ? this->correctedAcc * -1 : this->correctedAcc;
+    this->correctedAcc = (sqrt(this->measuredAccX * this->measuredAccX +
+                            this->measuredAccY * this->measuredAccY +
+                            this->measuredAccZ * this->measuredAccZ) 
+                            - this->idleAcc)
+                            * (this->measuredAccY >= 0 ? 1 : -1);
 
     // Check for sudden bumps
-    if (abs(this->correctedAcc - this->prevCorrectedAcc) > BUMP_THRESHOLD) {
+    if (abs(this->prevAcc - this->correctedAcc) > BUMP_THRESHOLD && FILTER_DELTA) {
         unsigned long currentTime = millis();
+        this->numBumps++;
 
-        // Check if the update timeout has elapsed
-        if (currentTime - this->lastUpdateTime > UPDATE_TIMEOUT) {
-            Serial.println("Acceleration update overdue; applying fallback."); // forcing changes
+        // Calculate the corrected acceleration 
+        if (this->correctedAcc < this->minBump) { 
+            this->minBump = this->correctedAcc; 
+        } 
+        if (this->correctedAcc > this->maxBump) { 
+            this->maxBump = this->correctedAcc; 
+        }
+
+        // Overrides if bumps are within a delta (therefore not a bump)
+        if (this->numBumps > SAMPLE_SIZE_BUMPS) {
+            if (this->maxBump - this->minBump < BUMP_THRESHOLD) {
+                Serial.println("Bumps seem consitent: Overriding");
+                this->smoothedAcc = this->correctedAcc;
+                this->prevAcc = this->correctedAcc;
+            }
+            this->minBump = 100000;
+            this->maxBump = -100000;
+            this->numBumps = 0;
+        }
+
+        // Forcing update if too much time has passed
+        else if (currentTime - this->lastUpdateTime > BUMP_OVERRIDE_TIME && currentTime - this->lastForceUpdate > MIN_TIME_BETWEEN_OVERRIDES) {
+            Serial.println("Too much time passed since last update: Overriding."); // forcing changes
             // Force an update with the current corrected value
-            this->smoothedAcc = this->smoothedAcc * (1 - SMOOTHING_FACTOR) + this->correctedAcc * SMOOTHING_FACTOR;
+            this->smoothedAcc = this->correctedAcc;
+            this->prevAcc = this->correctedAcc;
             this->lastUpdateTime = currentTime; // Reset the last update time
+            this->lastForceUpdate = currentTime;
+            this->minBump = 100000;
+            this->maxBump = -100000;
+            this->numBumps = 0;
         } else {
             // Skip the update since its a bump
-            
+            Serial.print("BUMP DETECTED: ");
+            Serial.println(this->correctedAcc);
             return;
         }
     }
 
-    this->accelerating = this->measuredAccY >= 0;
+    if (!FILTER_AVG) {
+        this->numSamples = AVG_SAMPLE_SIZE;
+        this->sumSamples = this->correctedAcc;
+        this->prevAcc = this->correctedAcc;
+    }
+    if (this->numSamples < AVG_SAMPLE_SIZE) {
+        this->sumSamples += this->correctedAcc;
+        this->numSamples++;
+        this->prevAcc = this->correctedAcc;
+        // Serial.print("Raw: ");
+        // Serial.println(this->correctedAcc);
+    }
+    else {
+        // Calculate the average acceleration
+        this->avgAcc = this->sumSamples / AVG_SAMPLE_SIZE;
+        // Serial.print("Avg: ");
+        // Serial.println(this->avgAcc);
 
-    // Smooth the acceleration
-    this->smoothedAcc = this->smoothedAcc * (1 - SMOOTHING_FACTOR) + this->correctedAcc * SMOOTHING_FACTOR;
+        this->smoothedAcc = this->smoothedAcc * (1 - SMOOTHING_FACTOR) + this->correctedAcc * SMOOTHING_FACTOR;
 
-    // Update previous corrected acceleration and timestamp
-    this->prevCorrectedAcc = this->correctedAcc;
-    lastUpdateTime = millis();
+        // Update previous corrected acceleration and timestamp and reset the sample data
+        this->numSamples = 0;
+        this->avgAcc = 0.0;
+        this->sumSamples = 0.0;
+
+        this->prevAcc = this->smoothedAcc;
+        lastUpdateTime = millis();
+        Serial.print("Acc: ");
+        Serial.println(this->smoothedAcc);
+    }
 }